worked on grocery list logic in service layer - test currently failing - probably my error

+++ java/src/main/java/com/techelevator/business/MealService.java
@@ -1,9 +1,11 @@
 package com.techelevator.business;
 
+import com.techelevator.exceptions.MealNotFoundException;
 import com.techelevator.model.Meal;
 import com.techelevator.model.MealPlan;
 import com.techelevator.model.User;
 import com.techelevator.repository.*;
+import com.techelevator.util.BasicLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,9 +26,8 @@ public class MealService {
     @Autowired
     MealPlanRepository mealPlanRepository;
 
-    //TODO create meal
-    public Meal createMeal (String name) {
 
+    public Meal createMeal (String name) {
         Meal meal = new Meal();
         meal.setMealName(name);
 
@@ -35,9 +36,17 @@ public class MealService {
         return meal;
     }
 
-    //TODO delete meal
+
     public void deleteMeal (Long mealId) {
-        mealRepository.delete(mealRepository.findByMealId(mealId));
+        try {
+            if (mealRepository.findByMealId(mealId) == null) {
+                throw new MealNotFoundException();
+            } else {
+                mealRepository.delete(mealRepository.findByMealId(mealId));
+            }
+        } catch (Exception e) {
+            BasicLogger.log(e.getMessage());
+        }
     }
 
     //TODO update meal
